+++ src/etheroll/roll_results.py
@@ -3,10 +3,10 @@ from kivy.clock import Clock, mainthread
 from kivy.properties import ListProperty
 from kivymd.label import MDLabel
 from kivymd.list import ILeftBody, ThreeLineAvatarListItem
+from pyetheroll.constants import ROUND_DIGITS
 
 from etheroll.ui_utils import Dialog, SubScreen, load_kv_from_py
 from etheroll.utils import run_in_thread
-from pyetheroll.constants import ROUND_DIGITS
 
 load_kv_from_py(__file__)